metadataExtractor/requirements.txt
newspaper3k
lxml
lxml_html_clean
fastapi
uvicorn
